this types the server connector in InMemotyClientConnector as ServerConnector as well as making it a final property one has to give in through the constructor. Due to this change several classes have to move to "combined" as they now reference a server class from the client side. Before this was hidden through the usage of Object
