+++ subprojects/combined/src/main/groovy/org/opendolphin/core/client/comm/SynchronousInMemoryClientConnector.groovy
@@ -18,7 +18,6 @@ package org.opendolphin.core.client.comm
 
 import groovy.transform.InheritConstructors
 import groovy.util.logging.Log
-import org.opendolphin.core.client.ClientDolphin
 import org.opendolphin.core.comm.Command
 
 /** An in-memory client connector without any asynchronous calls such that
@@ -29,10 +28,6 @@ import org.opendolphin.core.comm.Command
 
 @Log @InheritConstructors
 class SynchronousInMemoryClientConnector extends InMemoryClientConnector {
-
-    SynchronousInMemoryClientConnector(ClientDolphin clientDolphin) {
-        super(clientDolphin)
-    }
 
     protected void startCommandProcessing() {
         /* do nothing! */
