+++ subprojects/combined/src/test/groovy/org/opendolphin/core/client/comm/InMemoryClientConnectorTests.groovy
@@ -2,37 +2,36 @@ package org.opendolphin.core.client.comm
 
 import org.opendolphin.core.client.ClientDolphin
 import org.opendolphin.core.comm.Command
-
-import java.util.concurrent.CountDownLatch
+import org.opendolphin.core.server.ServerConnector
 
 class InMemoryClientConnectorTests extends GroovyTestCase {
 
     void testCallConnector_NoServerConnectorWired() {
-        InMemoryClientConnector connector = new InMemoryClientConnector(new ClientDolphin())
+        InMemoryClientConnector connector = new InMemoryClientConnector(new ClientDolphin(), null)
         assert [] == connector.transmit([new Command()])
     }
 
     void testCallConnector_ServerWired() {
-        InMemoryClientConnector connector = new InMemoryClientConnector(new ClientDolphin())
-        def command = new Command()
         boolean serverCalled = false
-        connector.serverConnector = [receive: { cmd ->
+        def serverConnector = [receive: { cmd ->
             serverCalled = true
             return []
-        }]
+        }] as ServerConnector
+        InMemoryClientConnector connector = new InMemoryClientConnector(new ClientDolphin(), serverConnector)
+        def command = new Command()
         connector.transmit([command])
         assert serverCalled
     }
 
     void testCallConnector_ServerWiredWithSleep() {
-        InMemoryClientConnector connector = new InMemoryClientConnector(new ClientDolphin())
-        connector.sleepMillis = 10
-        def command = new Command()
         boolean serverCalled = false
-        connector.serverConnector = [receive: { cmd ->
+        def serverConnector = [receive: { cmd ->
             serverCalled = true
             return []
-        }]
+        }] as ServerConnector
+        InMemoryClientConnector connector = new InMemoryClientConnector(new ClientDolphin(), serverConnector)
+        connector.sleepMillis = 10
+        def command = new Command()
         connector.transmit([command])
         assert serverCalled
     }
